+++ tests/Main.fs
@@ -102,16 +102,32 @@ let initialDatabase =
 [<Import("*", "json-server")>]
 let jsonServer : obj = jsNative
 
+[<Import("default", "multer")>]
+let multer : (obj -> obj) = jsNative
+
 [<Import("default", "./fake-delete.js")>]
 let fakeDeleteHandler : obj = jsNative
 
 [<Import("default", "./fake-unit.js")>]
 let fakeUnitHandler : obj = jsNative
 
+[<Import("default", "./fake-form-data.js")>]
+let fakeFormDataHandler : (obj -> obj) = jsNative
+
 [<Import("default", "./fake-error-report.js")>]
 let fakeErrorReportHandler : obj = jsNative
 
 Node.Api.``global``?fetch <- import "*" "node-fetch"
+Node.Api.``global``?Blob <- importDefault "fetch-blob"
+Node.Api.``global``?FormData <- importDefault "form-data"
+
+//FIXME: this looks like a bug in formdata-node or node-blob
+//       we need to update node-fetch to version-3
+//       ref: https://github.com/form-data/form-data/issues/220
+//       ref: https://github.com/form-data/form-data/issues/359
+[<Import("default", "./_bugfix_.js")>]
+let monkeyPatch : (obj -> unit) = jsNative
+monkeyPatch()
 
 describe "Thoth.Fetch" <| fun _ ->
 
@@ -136,6 +152,9 @@ describe "Thoth.Fetch" <| fun _ ->
                 "logger" ==> false
             ]
 
+        let upload = multer(createObj [ "storage" ==> multer?memoryStorage() ])
+        let fakeFormDataHandler = fakeFormDataHandler(databaseCreationDate)
+
         server?``use``(jsonServer?defaults(defaultOptions))
         server?delete("/fake-delete", fakeDeleteHandler)
         server?``get``("/get/unit", fakeUnitHandler)
@@ -143,6 +162,12 @@ describe "Thoth.Fetch" <| fun _ ->
         server?delete("/delete/unit", fakeUnitHandler)
         server?put("/put/unit", fakeUnitHandler)
         server?patch("/patch/unit", fakeUnitHandler)
+        server?put("/fake-form-data/book", upload?any(), fakeFormDataHandler)
+        server?post("/fake-form-data/book", upload?any(), fakeFormDataHandler)
+        server?patch("/fake-form-data/book", upload?any(), fakeFormDataHandler)
+        server?put("/fake-form-data/author", upload?any(), fakeFormDataHandler)
+        server?post("/fake-form-data/author", upload?any(), fakeFormDataHandler)
+        server?patch("/fake-form-data/author", upload?any(), fakeFormDataHandler)
         server?``get``("/get/fake-error-report", fakeErrorReportHandler)
         server?``use``(jsonServer?router(dbFile))
         serverInstance <- server?listen(3000, !!ignore)
@@ -151,6 +176,8 @@ describe "Thoth.Fetch" <| fun _ ->
         serverInstance?close()
 
     // End of the set up
+
+
 
     describe "Fetch.fetchAs" <| fun _ ->
 
@@ -423,6 +450,8 @@ Expecting a datetime but instead got: undefined
             |> Promise.catch d
             |> Promise.start
 
+
+
     describe "Fetch.get" <| fun _ ->
         it "Fetch.get works with manual decoder" <| fun d ->
             promise {
@@ -617,6 +646,8 @@ Expecting a datetime but instead got: undefined
                 d()
             } |> Promise.catch d
             |> Promise.start
+
+
 
     describe "Fetch.post" <| fun _ ->
         it "Fetch.post works with manual coder" <| fun d ->
@@ -1078,6 +1109,8 @@ Expecting a datetime but instead got: undefined
             } |> Promise.catch d
             |> Promise.start
 
+
+
     describe "Fetch.patch" <| fun _ ->
         it "Fetch.patch works with manual coder" <| fun d ->
             promise {
@@ -1282,6 +1315,8 @@ Expecting a datetime but instead got: undefined
             } |> Promise.catch d
             |> Promise.start
 
+
+
     describe "Fetch.delete" <| fun _ ->
 
         it "Fetch.detele can be just simple" <| fun d ->
@@ -1437,6 +1472,729 @@ Expecting a datetime but instead got: undefined
                 d()
             } |> Promise.catch d
             |> Promise.start
+
+
+
+
+    describe "Fetch.tryPostAsFormData" <| fun _ ->
+        it "Fetch.tryPostAsFormData works with manual coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPost(formData, "http://localhost:3000/fake-form-data/book", decoder = Book.Decoder)
+                let expected = Ok updatedBook
+
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPostAsFormData works with extra coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPost(formData, "http://localhost:3000/fake-form-data/book", extra = bookCoder)
+                let expected = Ok updatedBook
+
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPostAsFormData works with auto coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPost(formData, "http://localhost:3000/fake-form-data/book", caseStrategy = CamelCase)
+                let expected = Ok updatedBook
+
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPostAsFormData throw an exception explaining why the extra coder failed" <| fun d ->
+            promise {
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPost<Author>(formData, "http://localhost:3000/fake-form-data/author", extra = brokenAuthorCoder, caseStrategy = CamelCase)
+                let expected =
+                    Error(
+                        DecodingFailed(
+                            """
+Error at: `$`
+Expecting an object with a field named `author` but instead got:
+{
+    "id": 1,
+    "name": "hello-world"
+}
+                        """.Trim()
+                    ))
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPostAsFormData throw an exception explaining why the auto decoder failed" <| fun d ->
+            promise {
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPost<Book>(formData, "http://localhost:3000/fake-form-data/author", caseStrategy = CamelCase)
+
+                let expected =
+                    Error(
+                        DecodingFailed(
+                            """
+Error at: `$.createdAt`
+Expecting a datetime but instead got: undefined
+                        """.Trim()
+                    ))
+
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPostAsFormData works with unit response" <| fun d ->
+            promise {
+                let! res = Fetch.tryPost<unit>(null :> Browser.Types.FormData, "http://localhost:3000/post/unit")
+                let expected = Ok ()
+                Assert.AreEqual(res, expected)
+                d()
+            } |> Promise.catch d
+            |> Promise.start
+
+    describe "Fetch.postAsFormData" <| fun _ ->
+        it "Fetch.postAsFormData works with manual coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.post(formData, "http://localhost:3000/fake-form-data/book", decoder = Book.Decoder)
+
+                Assert.AreEqual(res, updatedBook)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.postAsFormData works with extra coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.post(formData, "http://localhost:3000/fake-form-data/book",  extra = bookCoder)
+
+                Assert.AreEqual(res, updatedBook)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.postAsFormData works with auto coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.post(formData, "http://localhost:3000/fake-form-data/book", caseStrategy = CamelCase)
+
+                Assert.AreEqual(res, updatedBook)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.postAsFormData throw an exception explaining why the extra coder failed" <| fun d ->
+            promise {
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! _ = Fetch.post<Author>(formData, "http://localhost:3000/fake-form-data/book", extra = brokenAuthorCoder, caseStrategy = CamelCase)
+                d()
+            }
+            |> Promise.catch (fun error ->
+                let expected =
+                    """
+[Thoth.Fetch] Error while decoding the response:
+
+Error at: `$`
+Expecting an object with a field named `author` but instead got:
+{
+    "id": 1,
+    "name": "hello-world"
+}
+                    """.Trim()
+                Assert.AreEqual(error.Message, expected)
+                d()
+            )
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.postAsFormData throw an exception explaining why the auto decoder failed" <| fun d ->
+            promise {
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! _ = Fetch.post<Book>(formData, "http://localhost:3000/fake-form-data/author", caseStrategy = CamelCase)
+                d()
+            }
+            |> Promise.catch (fun error ->
+                let expected =
+                    """
+[Thoth.Fetch] Error while decoding the response:
+
+Error at: `$.createdAt`
+Expecting a datetime but instead got: undefined
+                    """.Trim()
+                Assert.AreEqual(error.Message, expected)
+                d()
+            )
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.postAsFormData works with unit response" <| fun d ->
+            promise {
+                let! res = Fetch.post<unit>(null :> Browser.Types.FormData, "http://localhost:3000/post/unit")
+                Assert.AreEqual(res, ())
+                d()
+            } |> Promise.catch d
+            |> Promise.start
+
+
+
+    describe "Fetch.tryPutAsFormData" <| fun _ ->
+        it "Fetch.tryPutAsFormData works with manual coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPut(formData, "http://localhost:3000/fake-form-data/book", decoder = Book.Decoder)
+                let expected = Ok updatedBook
+
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPutAsFormData works with extra coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPut(formData, "http://localhost:3000/fake-form-data/book", extra = bookCoder)
+                let expected = Ok updatedBook
+
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPutAsFormData works with auto coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPut(formData, "http://localhost:3000/fake-form-data/book", caseStrategy = CamelCase)
+                let expected = Ok updatedBook
+
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPutAsFormData throw an exception explaining why the extra coder failed" <| fun d ->
+            promise {
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPut<Author>(formData, "http://localhost:3000/fake-form-data/author", extra = brokenAuthorCoder, caseStrategy = CamelCase)
+                let expected =
+                    Error(
+                        DecodingFailed(
+                            """
+Error at: `$`
+Expecting an object with a field named `author` but instead got:
+{
+    "id": 1,
+    "name": "hello-world"
+}
+                        """.Trim()
+                    ))
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPutAsFormData throw an exception explaining why the auto decoder failed" <| fun d ->
+            promise {
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPut<Book>(formData, "http://localhost:3000/fake-form-data/author", caseStrategy = CamelCase)
+
+                let expected =
+                    Error(
+                        DecodingFailed(
+                            """
+Error at: `$.createdAt`
+Expecting a datetime but instead got: undefined
+                        """.Trim()
+                    ))
+
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPutAsFormData works with unit response" <| fun d ->
+            promise {
+                let! res = Fetch.tryPut<unit>(null :> Browser.Types.FormData, "http://localhost:3000/put/unit")
+                let expected = Ok ()
+                Assert.AreEqual(res, expected)
+                d()
+            } |> Promise.catch d
+            |> Promise.start
+
+    describe "Fetch.putAsFormData" <| fun _ ->
+        it "Fetch.putAsFormData works with manual coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.put(formData, "http://localhost:3000/fake-form-data/book", decoder = Book.Decoder)
+
+                Assert.AreEqual(res, updatedBook)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.putAsFormData works with extra coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.put(formData, "http://localhost:3000/fake-form-data/book",  extra = bookCoder)
+
+                Assert.AreEqual(res, updatedBook)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.putAsFormData works with auto coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.put(formData, "http://localhost:3000/fake-form-data/book", caseStrategy = CamelCase)
+
+                Assert.AreEqual(res, updatedBook)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.putAsFormData throw an exception explaining why the extra coder failed" <| fun d ->
+            promise {
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! _ = Fetch.put<Author>(formData, "http://localhost:3000/fake-form-data/book", extra = brokenAuthorCoder, caseStrategy = CamelCase)
+                d()
+            }
+            |> Promise.catch (fun error ->
+                let expected =
+                    """
+[Thoth.Fetch] Error while decoding the response:
+
+Error at: `$`
+Expecting an object with a field named `author` but instead got:
+{
+    "id": 1,
+    "name": "hello-world"
+}
+                    """.Trim()
+                Assert.AreEqual(error.Message, expected)
+                d()
+            )
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.putAsFormData throw an exception explaining why the auto decoder failed" <| fun d ->
+            promise {
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! _ = Fetch.put<Book>(formData, "http://localhost:3000/fake-form-data/author", caseStrategy = CamelCase)
+                d()
+            }
+            |> Promise.catch (fun error ->
+                let expected =
+                    """
+[Thoth.Fetch] Error while decoding the response:
+
+Error at: `$.createdAt`
+Expecting a datetime but instead got: undefined
+                    """.Trim()
+                Assert.AreEqual(error.Message, expected)
+                d()
+            )
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.putAsFormData works with unit response" <| fun d ->
+            promise {
+                let! res = Fetch.put<unit>(null :> Browser.Types.FormData, "http://localhost:3000/put/unit")
+                Assert.AreEqual(res, ())
+                d()
+            } |> Promise.catch d
+            |> Promise.start
+
+
+
+    describe "Fetch.tryPatchAsFormData" <| fun _ ->
+        it "Fetch.tryPatchAsFormData works with manual coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPatch(formData, "http://localhost:3000/fake-form-data/book", decoder = Book.Decoder)
+                let expected = Ok updatedBook
+
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPatchAsFormData works with extra coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPatch(formData, "http://localhost:3000/fake-form-data/book", extra = bookCoder)
+                let expected = Ok updatedBook
+
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPatchAsFormData works with auto coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPatch(formData, "http://localhost:3000/fake-form-data/book", caseStrategy = CamelCase)
+                let expected = Ok updatedBook
+
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPatchAsFormData throw an exception explaining why the extra coder failed" <| fun d ->
+            promise {
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPatch<Author>(formData, "http://localhost:3000/fake-form-data/author", extra = brokenAuthorCoder, caseStrategy = CamelCase)
+                let expected =
+                    Error(
+                        DecodingFailed(
+                            """
+Error at: `$`
+Expecting an object with a field named `author` but instead got:
+{
+    "id": 1,
+    "name": "hello-world"
+}
+                        """.Trim()
+                    ))
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPatchAsFormData throw an exception explaining why the auto decoder failed" <| fun d ->
+            promise {
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.tryPatch<Book>(formData, "http://localhost:3000/fake-form-data/author", caseStrategy = CamelCase)
+
+                let expected =
+                    Error(
+                        DecodingFailed(
+                            """
+Error at: `$.createdAt`
+Expecting a datetime but instead got: undefined
+                        """.Trim()
+                    ))
+
+                Assert.AreEqual(res, expected)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.tryPatchAsFormData works with unit response" <| fun d ->
+            promise {
+                let! res = Fetch.tryPatch<unit>(null :> Browser.Types.FormData, "http://localhost:3000/patch/unit")
+                let expected = Ok ()
+                Assert.AreEqual(res, expected)
+                d()
+            } |> Promise.catch d
+            |> Promise.start
+
+    describe "Fetch.patchAsFormData" <| fun _ ->
+        it "Fetch.patchAsFormData works with manual coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.patch(formData, "http://localhost:3000/fake-form-data/book", decoder = Book.Decoder)
+
+                Assert.AreEqual(res, updatedBook)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.patchAsFormData works with extra coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.patch(formData, "http://localhost:3000/fake-form-data/book",  extra = bookCoder)
+
+                Assert.AreEqual(res, updatedBook)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.patchAsFormData works with auto coder" <| fun d ->
+            promise {
+                let! originalBook = Fetch.fetchAs("http://localhost:3000/books/1", caseStrategy = CamelCase)
+                let updatedBook =
+                    { originalBook with Title = "hello-world" }
+
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! res = Fetch.patch(formData, "http://localhost:3000/fake-form-data/book", caseStrategy = CamelCase)
+
+                Assert.AreEqual(res, updatedBook)
+                d()
+            }
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.patchAsFormData throw an exception explaining why the extra coder failed" <| fun d ->
+            promise {
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! _ = Fetch.patch<Author>(formData, "http://localhost:3000/fake-form-data/book", extra = brokenAuthorCoder, caseStrategy = CamelCase)
+                d()
+            }
+            |> Promise.catch (fun error ->
+                let expected =
+                    """
+[Thoth.Fetch] Error while decoding the response:
+
+Error at: `$`
+Expecting an object with a field named `author` but instead got:
+{
+    "id": 1,
+    "name": "hello-world"
+}
+                    """.Trim()
+                Assert.AreEqual(error.Message, expected)
+                d()
+            )
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.patchAsFormData throw an exception explaining why the auto decoder failed" <| fun d ->
+            promise {
+                let bytes = Fable.Core.JS.Constructors.Uint8Array.Create ( System.Text.Encoding.UTF8.GetBytes("hello-world") )
+                let file = Browser.Blob.Blob.Create([| bytes |])
+                let formData = Browser.Blob.FormData.Create()
+                formData.append("testField", file, "test.txt")
+
+                let! _ = Fetch.patch<Book>(formData, "http://localhost:3000/fake-form-data/author", caseStrategy = CamelCase)
+                d()
+            }
+            |> Promise.catch (fun error ->
+                let expected =
+                    """
+[Thoth.Fetch] Error while decoding the response:
+
+Error at: `$.createdAt`
+Expecting a datetime but instead got: undefined
+                    """.Trim()
+                Assert.AreEqual(error.Message, expected)
+                d()
+            )
+            |> Promise.catch d
+            |> Promise.start
+
+        it "Fetch.patchAsFormData works with unit response" <| fun d ->
+            promise {
+                let! res = Fetch.patch<unit>(null :> Browser.Types.FormData, "http://localhost:3000/patch/unit")
+                Assert.AreEqual(res, ())
+                d()
+            } |> Promise.catch d
+            |> Promise.start
+
+
 
     describe "Errors" <| fun _ ->
         it "A 404 should be reported as Bad Status" <| fun d ->